+++ apps/client/src/app/components/login-with-access-token-dialog/login-with-access-token-dialog.component.ts
@@ -1,3 +1,4 @@
+import { GfDialogHeaderComponent } from '@ghostfolio/client/components/dialog-header/dialog-header.component';
 import { InternetIdentityService } from '@ghostfolio/client/services/internet-identity.service';
 import {
   KEY_STAY_SIGNED_IN,
@@ -5,22 +6,44 @@ import {
 } from '@ghostfolio/client/services/settings-storage.service';
 import { TokenStorageService } from '@ghostfolio/client/services/token-storage.service';
 
+import { CommonModule } from '@angular/common';
 import { ChangeDetectionStrategy, Component, Inject } from '@angular/core';
-import { FormControl, Validators } from '@angular/forms';
-import { MatCheckboxChange } from '@angular/material/checkbox';
-import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
+import { FormControl, ReactiveFormsModule, Validators } from '@angular/forms';
+import { MatButtonModule } from '@angular/material/button';
+import {
+  MatCheckboxChange,
+  MatCheckboxModule
+} from '@angular/material/checkbox';
+import {
+  MAT_DIALOG_DATA,
+  MatDialogModule,
+  MatDialogRef
+} from '@angular/material/dialog';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatInputModule } from '@angular/material/input';
 import { Router } from '@angular/router';
+import { IonIcon } from '@ionic/angular/standalone';
 import { addIcons } from 'ionicons';
 import { eyeOffOutline, eyeOutline } from 'ionicons/icons';
 
 @Component({
-  selector: 'gf-login-with-access-token-dialog',
   changeDetection: ChangeDetectionStrategy.OnPush,
+  imports: [
+    CommonModule,
+    GfDialogHeaderComponent,
+    IonIcon,
+    MatButtonModule,
+    MatCheckboxModule,
+    MatDialogModule,
+    MatFormFieldModule,
+    MatInputModule,
+    ReactiveFormsModule
+  ],
+  selector: 'gf-login-with-access-token-dialog',
   styleUrls: ['./login-with-access-token-dialog.scss'],
-  templateUrl: 'login-with-access-token-dialog.html',
-  standalone: false
+  templateUrl: './login-with-access-token-dialog.html'
 })
-export class LoginWithAccessTokenDialog {
+export class GfLoginWithAccessTokenDialogComponent {
   public accessTokenFormControl = new FormControl(
     this.data.accessToken,
     Validators.required
@@ -29,7 +52,7 @@ export class LoginWithAccessTokenDialog {
 
   public constructor(
     @Inject(MAT_DIALOG_DATA) public data: any,
-    public dialogRef: MatDialogRef<LoginWithAccessTokenDialog>,
+    public dialogRef: MatDialogRef<GfLoginWithAccessTokenDialogComponent>,
     private internetIdentityService: InternetIdentityService,
     private router: Router,
     private settingsStorageService: SettingsStorageService,