+++ apps/api/src/app/endpoints/sitemap/sitemap.controller.ts
@@ -8,8 +8,8 @@ import {
 import { Controller, Get, Res, VERSION_NEUTRAL, Version } from '@nestjs/common';
 import { format } from 'date-fns';
 import { Response } from 'express';
-import { readFileSync } from 'fs';
-import { join } from 'path';
+import { readFileSync } from 'node:fs';
+import { join } from 'node:path';
 
 import { SitemapService } from './sitemap.service';