+++ apps/api/src/app/portfolio/calculator/portfolio-calculator-test-utils.ts
@@ -1,4 +1,4 @@
-import { readFileSync } from 'fs';
+import { readFileSync } from 'node:fs';
 
 export const activityDummyData = {
   accountId: undefined,