+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Refactored the login with access token dialog component to standalone
 - Prefixed the `crypto`, `fs` and `path` imports with `node:`
 
 ## 2.198.0 - 2025-09-11