+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- Moved the holdings table into the holdings section on the public page
+- Refactored the login with access token dialog component to standalone
+
 ## 2.198.0 - 2025-09-11
 
 ### Changed