+++ CHANGELOG.md
@@ -10,13 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Moved the holdings table into the holdings section on the public page
+- Migrated to the _Prisma Configuration File_ approach (`prisma.config.ts`)
 - Refactored the login with access token dialog component to standalone
 
 ## 2.198.0 - 2025-09-11
 
 ### Changed
 
-- Migrated to Prisma Configuration File `prisma.config.ts`
 - Extended the variations of the interstitials for the subscription
 - Renamed the job identifier column in the jobs queue view of the admin control panel
 - Refactored the markets page to standalone