+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Improved the search in the _Yahoo Finance_ service
 - Moved the holdings table into the holdings section on the public page
 - Refactored the login with access token dialog component to standalone