+++ CHANGELOG.md
@@ -302,6 +302,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Set up the language localization for the static portfolio analysis rule: _Economic Market Cluster Risks_ (Developed Markets)
 - Set up the language localization for the static portfolio analysis rule: _Economic Market Cluster Risks_ (Emerging Markets)
+- Added stories for the entity logo image component
 
 ### Changed