+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improved the search in the _Yahoo Finance_ service
 - Moved the holdings table into the holdings section on the public page
 - Refactored the login with access token dialog component to standalone
+- Prefixed the `crypto`, `fs` and `path` imports with `node:`
 
 ## 2.198.0 - 2025-09-11